\name{caspar-package}
\alias{caspar-package}
\alias{caspar}
\docType{package}
\title{
CaSpaR: Clustered and sparse regression.
}
\description{
Caspar is an algorithm for building sparse structured linear regression models.  The method is particularly adapted to the setting in which the structure can be defined as a pairwise distance between predictors.  caspar.main() is the key function of this package.  See demo.caspar() and demo.hiv() for demonstrations.
}
\details{
\tabular{ll}{
Package: \tab caspar\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-29\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr}

}
\author{
Daniel Percival

Maintainer: Daniel Percival <dperciva@andrew.cmu.edu>
%~~ The author and/or maintainer of the package ~~
}
\references{
Structured, Sparse Regression With Application to HIV Drug Resistance.  AOAS, to appear.  See: http://arxiv.org/abs/1002.3128.

Rhee, S.-Y., Gonzales, M. J., Kantor, R., Betts, B. J., Ravela, J. and 
Shafer, R. W. (2003). 
  Human immunodeficiency virus reverse transcriptase and protease sequence database. 
  \emph{Nucleic Acids Research 31: 298-303}.
}
\keyword{ package }
\seealso{
}
\examples{
}
