\name{HIVDAT7SQV}
\alias{HIVDAT7SQV}
\title{HIV Data: Drug SQV}
\usage{data(HIVDAT7SQV)}
\description{HIV Data for protease inhibitor drug SQV.}
\references{
Rhee, S.-Y., Gonzales, M. J., Kantor, R., Betts, B. J., Ravela, J. and 
Shafer, R. W. (2003). 
  Human immunodeficiency virus reverse transcriptase and protease sequence database. 
  \emph{Nucleic Acids Research 31: 298-303}. 
}
\keyword{datasets}
\seealso{
\code{\link{demo.hiv}}
}
